extract timezone handling from event into concern
